requirements.txt
torch
torchvision
pandas
umap-learn
openslide-python
h5py
scikit-learn
seaborn
matplotlib
Pillow
kde
scipy
scikit-image
pyvips
webcolors
lifelines
scikit-survival
openpyxl
flask
opencv-python==4.5.5.64



















